requirements.txt
boto==2.48.0
boto3==1.7.4
botocore==1.10.10
click==6.7
Flask==0.12.2
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
imgaug==0.2.5
Jinja2==2.10
Keras==2.1.5
matplotlib==2.2.2
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
requests==2.18.4
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
tqdm==4.23.4
urllib3==1.22
uWSGI==2.0.17
Werkzeug==0.14.1
